backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
